Remove `/tmp` directory for artifacts
On MacOS, this results in:

```
Error: Provided rootDirectory /tmp is not a valid directory
```

On Windows this results in:

```
Error: No files were found with the provided path: /tmp/gotest.windows.log. No artifacts will be uploaded.
```

Both seem to be unhappy with `/tmp`, since the workspace is implicitly cleaned up at the end of the workflow, there shouldn't be a need to specify `/tmp`.

Signed-off-by: Yuri Norwood <106889957+norwd@users.noreply.github.com>